Moved utility functions from __init__.py to common.py
