+++ exasol_script_languages_container_ci/lib/ci_build.py
@@ -4,7 +4,7 @@ import click
 from exasol_integration_test_docker_environment.cli.commands.build_test_container import build_test_container
 from exasol_script_languages_container_tool.cli.commands import build
 
-from exasol_script_languages_container_ci.lib import print_docker_images
+from exasol_script_languages_container_ci.lib.common import print_docker_images
 import logging
