+++ exasol_script_languages_container_ci/lib/ci_test.py
@@ -5,7 +5,7 @@ import click
 from exasol_script_languages_container_tool.cli.commands import run_db_test
 
 import exasol_script_languages_container_ci
-from exasol_script_languages_container_ci.lib import print_docker_images
+from exasol_script_languages_container_ci.lib.common import print_docker_images
 
 
 def ci_test(ctx: click.Context, flavor_path: Tuple[str, ...]):
@@ -13,7 +13,7 @@ def ci_test(ctx: click.Context, flavor_path: Tuple[str, ...]):
     Run db tests
     """
 
-    if "skip tests" not in exasol_script_languages_container_ci.lib.get_last_commit_message():
+    if "skip tests" not in exasol_script_languages_container_ci.lib.common.get_last_commit_message():
         logging.info(f"Running command 'run_db_test' with parameters {locals()}")
         ctx.invoke(run_db_test, flavor_path=flavor_path, workers=7)
         logging.info(f"Running command 'run_db_test' for linker_namespace_sanity with parameters {locals()}")
